requirements.txt
Django~=4.1.5
djangorestframework~=3.14.0